Apply review-findings triage: int64 wire-byte estimates, doc comments, cache-race test
- recordEstimateBytes/recordBatchEstimateBytes return int64 so the per-record
  size gate can't be bypassed by int32 overflow on a pathologically large
  record; buffer wire-byte accounting widened to match. Add a MaxBatchBytes
  sanity ceiling to Config.Validate.
- Document intentional designs raised in review: ClusterRecordBuffer's buffered
  counters are observability-only (callers own backpressure); ThrottleMillis is
  propagated but not enforced; name the golden-ratio prime in hashTopicPartition.
- Add a regression test for the CachedAgentStatsTracker purge-vs-compute race.
- Remove REVIEW_FINDINGS.md (superseded by the agreed triage).

+++ pkg/wgo/agent_record_buffer_test.go
@@ -366,7 +366,7 @@ func TestAgentRecordBuffer_BufferedWireBytes(t *testing.T) {
 			a.mu.Unlock()
 
 			actual := actualUncompressedMultiRecordBatchWireSize(tc.records)
-			assert.Equal(t, actual, running)
+			assert.Equal(t, int64(actual), running)
 		})
 	}
 }
@@ -426,7 +426,7 @@ func TestAgentRecordBuffer_BufferedWireBytes_AfterEarlyFlush(t *testing.T) {
 	a.mu.Unlock()
 
 	assert.Equal(t, second[0].Timestamp.UnixMilli(), anchor)
-	assert.Equal(t, actualUncompressedMultiRecordBatchWireSize(second), running)
+	assert.Equal(t, int64(actualUncompressedMultiRecordBatchWireSize(second)), running)
 }
 
 func TestAgentRecordBuffer_Close(t *testing.T) {
